requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
Flask-Compress==1.14
flask-sqlalchemy>=2.5.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
pyxlsb==1.0.10
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
xlwings==0.33.9
yarl==1.18.3
gunicorn==20.1.0
psycopg2-binary==2.9.9
gevent==25.5.1
gevent-websocket==0.10.1
redis>=4.5.0
python-dotenv==1.0.0
sqlalchemy>=1.4.0
